+++ patches/src/main/kotlin/dev/jkcarino/revanced/patches/reddit/layout/screenshotbanner/DisableScreenshotBannerPatch.kt
@@ -4,6 +4,7 @@ import app.revanced.patcher.extensions.InstructionExtensions.getInstruction
 import app.revanced.patcher.extensions.InstructionExtensions.replaceInstruction
 import app.revanced.patcher.patch.bytecodePatch
 import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
+import dev.jkcarino.revanced.patches.reddit.misc.firebase.spoofCertificateHashPatch
 import dev.jkcarino.revanced.util.returnEarly
 
 @Suppress("unused")
@@ -12,6 +13,8 @@ val disableScreenshotBannerPatch = bytecodePatch(
     description = "Disables the banner that shows up after taking a screenshot.",
     use = true,
 ) {
+    dependsOn(spoofCertificateHashPatch)
+
     compatibleWith("com.reddit.frontpage")
 
     execute {